% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{export_as_rpolars_df}
\alias{export_as_rpolars_df}
\title{export rpolars DataFrame from rust to rpolars}
\usage{
export_as_rpolars_df()
}
\value{
r-polars DataFrame
}
\description{
export rpolars DataFrame from rust to rpolars
}
\details{
this function requires r-polars polars package >= 0.8.0.
}
