% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{import_rpolars_df_and_print}
\alias{import_rpolars_df_and_print}
\title{import rpolars DataFrame to rust and print back to R terminal}
\usage{
import_rpolars_df_and_print(rdf)
}
\arguments{
\item{rdf}{r-polars DataFrame}
}
\description{
import rpolars DataFrame to rust and print back to R terminal
}
\details{
this function requires r-polars polars package >= 0.8.0
}
