% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{make_df}
\alias{make_df}
\title{show case zero copy export from rust-polars to r-polars}
\usage{
make_df()
}
\description{
show case zero copy export from rust-polars to r-polars
}
\details{
conversion via arrow stream such to ensure version compatability
}
