% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{test_round_trip}
\alias{test_round_trip}
\title{import rpolars DataFrame to rust and export again}
\usage{
test_round_trip(rdf)
}
\arguments{
\item{rdf}{rpolars dataframe}
}
\value{
rpolars dataframe
}
\description{
import rpolars DataFrame to rust and export again
}
\details{
this function requires r-polars polars package >= 0.8.0
}
